chore: include Vuetify date options
